backend/requirements.txt
alembic==1.13.1
annotated-types==0.7.0
asgiref==3.8.1
astroid==3.0.1
Babel==2.14.0
blinker==1.7.0
catalogue==2.0.10
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.1
cloudpathlib==0.20.0
confection==0.1.5
cycler==0.12.1
cymem==2.0.8
dill==0.3.7
Django==5.2
django-cors-headers==4.3.1
django-filter==25.1
djangorestframework==3.15.0
djangorestframework-simplejwt
fonttools==4.54.1
google-auth
google-auth-oauthlib
greenlet==3.0.3
idna==3.2
isort==5.12.0
itsdangerous==2.0.1
joblib==1.4.2
kiwisolver==1.4.7
langcodes==3.5.0
language_data==1.3.0
Mako==1.3.2
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==2.0.1
mccabe==0.7.0
mdurl==0.1.2
multipledispatch==1.0.0
murmurhash==1.0.10
nltk==3.9.1
packaging==24.0
passlib==1.7.4
pillow==10.2.0
platformdirs==4.0.0
preshed==3.0.9
psycopg2-binary
pydantic==2.10.0
pydantic_core==2.27.0
Pygments==2.18.0
pylint==3.0.2
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.1
regex==2024.11.6
requests==2.26.0
rich==13.9.4
setuptools==75.6.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
speaklater==1.3
sqlparse==0.5.3
srsly==2.4.8
threadpoolctl==3.5.0
toml==0.10.2
tomlkit==0.12.3
tqdm==4.67.0
typer==0.13.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.18
wasabi==1.1.3
weasel==0.4.1
wrapt==1.16.0
pytest
pytest-django
